+++ modules/tinymce/src/themes/silver/demo/ts/components/DemoHelpers.ts
@@ -116,13 +116,6 @@ const setupDemo = () => {
             hotspot: sink
           };
         },
-        toolbarOverflow: () => {
-          // NOTE: Non-sensical
-          return {
-            anchor: 'hotspot',
-            hotspot: sink
-          };
-        },
         banner: () => {
           // NOTE: Non-sensical
           return {