+++ modules/tinymce/src/themes/silver/main/ts/backstage/Anchors.ts
@@ -1,8 +1,8 @@
-import { AlloyComponent, Bubble, HotspotAnchorSpec, Layout, LayoutInside, MaxHeight, MaxWidth, NodeAnchorSpec, SelectionAnchorSpec } from '@ephox/alloy';
+import { AlloyComponent, Bubble, HotspotAnchorSpec, Layout, LayoutInside, MaxHeight, NodeAnchorSpec, SelectionAnchorSpec } from '@ephox/alloy';
 import { Option } from '@ephox/katamari';
 import { Element, Selection } from '@ephox/sugar';
 import Editor from 'tinymce/core/api/Editor';
-import { useFixedContainer, isToolbarLocationTop } from '../api/Settings';
+import { useFixedContainer } from '../api/Settings';
 
 const bubbleAlignments = {
   valignCentre: [],
@@ -73,20 +73,6 @@ const getBannerAnchor = (bodyElement: () => Element, lazyAnchorbar: () => AlloyC
   return useFixedToolbarContainer ? fixedToolbarAnchor : standardAnchor;
 };
 
-const getToolbarOverflowAnchor = (lazyMoreButton: () => AlloyComponent, isToolbarTop: boolean) => (): HotspotAnchorSpec => {
-  return {
-    anchor: 'hotspot',
-    hotspot: lazyMoreButton(),
-    overrides: {
-      maxWidthFunction: MaxWidth.expandable()
-    },
-    layouts: {
-      onRtl: () => isToolbarTop ? [ Layout.southeast, Layout.southwest ] : [ Layout.northeast, Layout.northwest ],
-      onLtr: () => isToolbarTop ? [ Layout.southwest, Layout.southeast ] : [ Layout.northwest, Layout.northeast ]
-    }
-  };
-};
-
 const getCursorAnchor = (editor: Editor, bodyElement: () => Element) => (): SelectionAnchorSpec => {
   return {
     anchor: 'selection',
@@ -108,14 +94,12 @@ const getNodeAnchor = (bodyElement) => (element: Option<Element>): NodeAnchorSpe
   };
 };
 
-const getAnchors = (editor: Editor, lazyAnchorbar: () => AlloyComponent, lazyMoreButton: () => AlloyComponent) => {
+const getAnchors = (editor: Editor, lazyAnchorbar: () => AlloyComponent) => {
   const useFixedToolbarContainer: boolean = useFixedContainer(editor);
   const bodyElement = (): Element => Element.fromDom(editor.getBody());
-  const isToolbarTop = isToolbarLocationTop(editor);
 
   return {
     toolbar: getToolbarAnchor(bodyElement, lazyAnchorbar, useFixedToolbarContainer),
-    toolbarOverflow: getToolbarOverflowAnchor(lazyMoreButton, isToolbarTop),
     banner: getBannerAnchor(bodyElement, lazyAnchorbar, useFixedToolbarContainer),
     cursor: getCursorAnchor(editor, bodyElement),
     node: getNodeAnchor(bodyElement)