+++ modules/alloy/changelog.md
@@ -6,11 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-# [5.1.0] - TBD
+# [6.0.0] - TBD
+
+### Removed
+- Removed `getMoreButton` and `getOverflow` methods from the `SplitSlidingToolbar` component.
+- Removed `getMoreButton` method from the `SplitFloatingToolbar` component.
+- Removed `getAnchor` configuration from the `SplitFloatingToolbar` component.
 
 ### Added
 - Added new `FloatingToolbar` component.
 - Added new vertical directionality for layouts
+- Added `setGroups` and `toggle` methods and `getBounds` configuration to the `FloatingToolbarButton` component.
  
 ### Fixed
 - Improved `Docking` scroll performance by only calculating the offset origin as required.